DjangoProject/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
